# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Control de Capital Humano API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # UDNs
  /udns:
    get:
      operationId: listUdns
      tags: [udns]
      summary: List all UDNs
      responses:
        "200":
          description: List of UDNs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Udn"
    post:
      operationId: createUdn
      tags: [udns]
      summary: Create a UDN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UdnInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Udn"
  /udns/{id}:
    get:
      operationId: getUdn
      tags: [udns]
      summary: Get UDN by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: UDN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Udn"
    patch:
      operationId: updateUdn
      tags: [udns]
      summary: Update UDN
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UdnUpdate"
      responses:
        "200":
          description: Updated UDN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Udn"
    delete:
      operationId: deleteUdn
      tags: [udns]
      summary: Delete UDN
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PROFILES
  /profiles:
    get:
      operationId: listProfiles
      tags: [profiles]
      summary: List all profiles
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"
    post:
      operationId: createProfile
      tags: [profiles]
      summary: Create a profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/{id}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get profile by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateProfile
      tags: [profiles]
      summary: Update profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    delete:
      operationId: deleteProfile
      tags: [profiles]
      summary: Delete profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ZONAS AUDITADAS
...
  /panels:
    get:
      operationId: listPanels
      tags: [panels]
      summary: List all panels
      responses:
        "200":
          description: List of panels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Panel"
    post:
      operationId: createPanel
      tags: [panels]
      summary: Create a panel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PanelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Panel"
  /panels/{id}:
    get:
      operationId: getPanel
      tags: [panels]
      summary: Get panel by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Panel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Panel"
    patch:
      operationId: updatePanel
      tags: [panels]
      summary: Update panel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PanelUpdate"
      responses:
        "200":
          description: Updated panel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Panel"
    delete:
      operationId: deletePanel
      tags: [panels]
      summary: Delete panel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # DEFECTOS
...
  /sides:
    get:
      operationId: listSides
      tags: [sides]
      summary: List all sides
      responses:
        "200":
          description: List of sides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Side"
    post:
      operationId: createSide
      tags: [sides]
      summary: Create a side
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SideInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Side"
  /sides/{id}:
    get:
      operationId: getSide
      tags: [sides]
      summary: Get side by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Side
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Side"
    patch:
      operationId: updateSide
      tags: [sides]
      summary: Update side
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SideUpdate"
      responses:
        "200":
          description: Updated side
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Side"
    delete:
      operationId: deleteSide
      tags: [sides]
      summary: Delete side
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ZONA VISUAL
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, puesto, area, udn_id, role, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        puesto:
          type: string
        area:
          type: string
        profile_id:
          type: ["integer", "null"]
        udn_id:
          type: ["integer", "null"]
        role:
          type: string
          enum: [superadmin, admin, user]
        created_at:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [name, email, password, puesto, area]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        puesto:
          type: string
        area:
          type: string
        profile_id:
          type: ["integer", "null"]
        udn_id:
          type: ["integer", "null"]
        role:
          type: string
          enum: [superadmin, admin, user]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        puesto:
          type: string
        area:
          type: string
        profile_id:
          type: ["integer", "null"]
        udn_id:
          type: ["integer", "null"]
        role:
          type: string
          enum: [superadmin, admin, user]

    # UDN
    Udn:
      type: object
      required: [id, name, code, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        description:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
    UdnInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
    UdnUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string

    # PROFILE
    Profile:
      type: object
      required: [id, name, permissions, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        permissions:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    ProfileInput:
      type: object
      required: [name, permissions]
      properties:
        name:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            type: string
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            type: string

    # ZONE
...
    Panel:
      type: object
      required: [id, name, columns, rows, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        diagram_url:
          type: ["string", "null"]
        columns:
          type: integer
          description: Number of columns (numbered)
        rows:
          type: integer
          description: Number of rows (lettered A, B, C...)
        zone_id:
          type: ["integer", "null"]
        side_id:
          type: ["integer", "null"]
        visual_zone_id:
          type: ["integer", "null"]
        alphanumeric_ids:
          type: array
          items:
            type: integer
          default: []
        created_at:
          type: string
          format: date-time
    PanelInput:
      type: object
      required: [name, columns, rows]
      properties:
        name:
          type: string
        description:
          type: string
        diagram_url:
          type: string
        columns:
          type: integer
        rows:
          type: integer
        zone_id:
          type: integer
        side_id:
          type: integer
        visual_zone_id:
          type: integer
        alphanumeric_ids:
          type: array
          items:
            type: integer
    PanelUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        diagram_url:
          type: string
        columns:
          type: integer
        rows:
          type: integer
        zone_id:
          type: integer
        side_id:
          type: integer
        visual_zone_id:
          type: integer
        alphanumeric_ids:
          type: array
          items:
            type: integer

    # AUDIT CAPTURE
...
    Side:
      type: object
      required: [id, name, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
    SideInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    SideUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    # VISUAL ZONE
...